requirements.txt
xlrd
openpyxl
pyyaml
pyinstaller